Second pass at exports conundrum
