+++ src/vs/workbench/api/browser/positron/mainThreadPlotsService.ts
@@ -6,8 +6,7 @@
 import { DisposableStore } from '../../../../base/common/lifecycle.js';
 import { ExtHostPlotsServiceShape, ExtHostPositronContext, MainPositronContext, MainThreadPlotsServiceShape } from '../../common/positron/extHost.positron.protocol.js';
 import { extHostNamedCustomer, IExtHostContext } from '../../../services/extensions/common/extHostCustomers.js';
-import { IPositronPlotsService } from '../../../services/positronPlots/common/positronPlots.js';
-import { PlotRenderSettings } from '../../common/positron/extHostTypes.positron.js';
+import { IPositronPlotsService, PlotRenderSettings } from '../../../services/positronPlots/common/positronPlots.js';
 
 @extHostNamedCustomer(MainPositronContext.MainThreadPlotsService)
 export class MainThreadPlotsService implements MainThreadPlotsServiceShape {
